fix security config, registration flow and service layering
AdSecurity — добавлена проверка доступа (автор объявления или ADMIN) для защиты бизнес-операций.
CommentSecurity — добавлена проверка доступа (автор комментария или ADMIN) для защиты редактирования/удаления.
WebSecurityConfig — настроен JdbcUserDetailsManager под таблицы БД, добавлены ролевые правила URL и открыт доступ к Swagger/регистрации.
AdvertisementController — убран X-User-Id, автор объявления теперь берется из Authentication.
AuthController — включена валидация входного запроса регистрации (@Valid).
CommentController — убран прямой доступ к UserRepository, контроллер работает через сервис.
UserController — усилена валидация RegisterUserRequest (вложенный user + пароль).
UserEntity — добавлено поле enabled для корректной работы Spring Security с БД.
CommentMapper — добавлен маппинг toEntity(...) для симметричного преобразования DTO↔Entity.
AdService — обновлен контракт создания объявления: вместо authorId используется authorEmail.
CommentService — обновлен контракт добавления комментария: вместо authorId используется authorEmail.
AdServiceImpl — автор объявления ищется по email текущего пользователя, сохранение идет через сервисный слой.
AuthServiceImpl — регистрация переведена на JPA-модель (users/roles) с хешированием пароля.
CommentServiceImpl — добавление комментария переведено на автора по email + использование CommentMapper.
UserServiceImpl — добавлены проверки дубля email и авто-создание роли USER при отсутствии.
Dockerfile — обновлены настройки контейнера (сборка/запуск приложения)(не обязательно, все равно мы не будем запускать).
pom.xml — обновлены зависимости/конфигурация сборки под текущую реализацию.
CommentRepository — обновлены методы доступа к комментариям для работы новых сервисных сценариев.

+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -6,6 +6,6 @@ import ru.skypro.homework.entity.CommentEntity;
 
 import java.util.List;
 @Repository
-public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
-    List<CommentEntity> findByAdId(Long adId);
+public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
+    List<CommentEntity> findByAdId(Integer adId);
 }
